Remove duplicate rational domain and cleanup code and semantic of methods
Duplication rational domain (CPRationalDom and CPRationalDomN) since one was used in CPFloatVar and CPDoubleVar as domain of error, and the other one was used as domain of value for CPRationalVar.
Now there is only one domain (with specific update methods for errors and values, depending on the variable calling it) and two notifier are specified on this domain (CPErrorVarNotifier and CPRationalVarNotifier).
CPRationalDom is the only rational domain to maintain now.